+++ frontend/onestopgo/src/categories/category-item.js
