+++ frontend/onestopgo/src/home/home-thunk.js
@@ -2,7 +2,7 @@ import {createAsyncThunk} from "@reduxjs/toolkit";
 import * as service from "./home-service"
 
 
-export const getHomeDataThunk = createAsyncThunk(
+export const getHomeDataThunk = createAsyncThunk (
     'onestopgo/homeData',
     async () => await service.getHomeData()
 )
